Separated the Authentication code from AppController.java to AuthUtils.java
