+++ dialogflow/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
 <string name="app_name">Dialogflow</string>
-<string name="gcp_project_id">GCP_PROJECT-ID</string>
+<string name="gcp_project_id">GCP_PROJECT_ID</string>
 <string name="knowledge_connectors">Knowledge Connectors</string>
 <string name="sentiment_analysis">Sentiment Analysis</string>
 <string name="text_to_speech">Text to speech</string>
